+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "behaveai"
+name = "behave-ai"
 version = "1.2.1"
 description = "Animal tracking and behaviour classification using YOLO and motion analysis"
 readme = "README.md"
@@ -19,11 +19,12 @@ dependencies = [
     "pyyaml",
     "opencv-python",
     "torch",
-    "torchvision"
+    "torchvision",
+    "click"
 ]
 
 [project.scripts]
-behaveai = "behaveai.BehaveAI:main"
+behaveai = "behaveai.cli:cli"
 
 [tool.hatch.build.targets.wheel]
 packages = ["behaveai"]
